feat: add index.html for contact redirection and .gitignore for macOS
Introduce an `index.html` file that serves as a contact page for Kotaro
Nishigori. The page includes a meta refresh tag to redirect users to
the specified Linktree URL after 5 seconds, providing a simple and
effective way to guide users to the contact's main link hub. The page
is styled to center content and provide a clean, minimalistic design.

Additionally, create a `.gitignore` file to exclude `.DS_Store` files,
which are automatically generated by macOS and are not relevant to the
project. This helps keep the repository clean from unnecessary files.